refactor and introduce SqlViewSelect
https://github.com/metasfresh/metasfresh-webui-api/issues/668
