+++ src/main/java/de/metas/ui/web/view/SqlViewRowIdsOrderedSelectionFactory.java
@@ -17,9 +17,9 @@ import com.google.common.base.Stopwatch;
 
 import de.metas.logging.LogManager;
 import de.metas.ui.web.document.filter.DocumentFilter;
+import de.metas.ui.web.view.descriptor.SqlAndParams;
 import de.metas.ui.web.view.descriptor.SqlViewBinding;
 import de.metas.ui.web.view.descriptor.SqlViewSelectionQueryBuilder;
-import de.metas.ui.web.view.descriptor.SqlViewSelectionQueryBuilder.SqlAndParams;
 import de.metas.ui.web.view.descriptor.SqlViewSelectionQueryBuilder.SqlCreateSelection;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.ui.web.window.datatypes.DocumentIdsSelection;
